move offsets.rs -> layout.rs
